requirements.txt
zeroconf
pynacl
aiosqlite
textual
pytest
